pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ec2-session-gate"
# NOTE: Version should match src/version.py (single source of truth)
# Update both files when changing version, or use: python -c "from src.version import __version__; print(__version__)"
version = "1.0.1"
description = "Cross-platform AWS SSM tunnel/connection manager for EC2 instances"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Walid Boussafa", email = "walid.boussafa@outlook.com"}
]
keywords = ["aws", "ec2", "ssm", "session-manager", "ssh", "rdp", "port-forwarding", "tunnel"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Flask>=2.3",
    "boto3>=1.34",
    "botocore>=1.34",
    "PyYAML>=6.0",
    "pywebview>=4.0",
    "cryptography>=41.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/boussaffawalid/ec2-session-gate"
Documentation = "https://github.com/boussaffawalid/ec2-session-gate#readme"
Repository = "https://github.com/boussaffawalid/ec2-session-gate"
Issues = "https://github.com/boussaffawalid/ec2-session-gate/issues"

[project.scripts]
ec2-session-gate = "run:main"
ec2-session-gate-gui = "run:main"

[tool.setuptools]
packages = ["src"]
py-modules = ["run"]

[tool.setuptools.package-data]
src = [
    "static/**/*",
    "static/templates/**/*",
    "defaults.json",
    "logging.yaml",
]
